Split Validation phase into validation/verify
