+++ klum-ast-runtime/src/main/java/com/blackbuild/klum/ast/util/ValidationPhase.java
@@ -23,59 +23,24 @@
  */
 package com.blackbuild.klum.ast.util;
 
-import com.blackbuild.groovy.configdsl.transform.Validate;
-import com.blackbuild.klum.ast.process.AbstractPhaseAction;
 import com.blackbuild.klum.ast.process.DefaultKlumPhase;
-import com.blackbuild.klum.ast.process.PhaseDriver;
-import com.blackbuild.klum.ast.util.layer3.ModelVisitor;
-import com.blackbuild.klum.ast.util.layer3.StructureUtil;
+import com.blackbuild.klum.ast.process.VisitingPhaseAction;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
-
-import java.util.ArrayList;
-import java.util.List;
 
 /**
  * Phase Action that validates the model.
  */
-public class ValidationPhase extends AbstractPhaseAction {
+public class ValidationPhase extends VisitingPhaseAction {
 
     public ValidationPhase() {
         super(DefaultKlumPhase.VALIDATE);
     }
 
     @Override
-    protected void doExecute() {
-        new Visitor().execute();
-    }
+    public void visit(@NotNull String path, @NotNull Object element, @Nullable Object container, @Nullable String nameOfFieldInContainer) {
+        if (KlumInstanceProxy.getProxyFor(element).getManualValidation()) return;
 
-    public static class Visitor implements ModelVisitor {
-
-        private final List<KlumValidationResult> aggregatedErrors = new ArrayList<>();
-        private Validate.Level currentMaxLevel = Validate.Level.NONE;
-
-        void execute() {
-            executeOn(PhaseDriver.getInstance().getRootObject(), Validator.getFailLevel());
-        }
-
-        List<KlumValidationResult> executeOn(Object root, Validate.Level failOnLevel) {
-            StructureUtil.visit(root, this);
-            if (currentMaxLevel.equalOrWorseThan(failOnLevel))
-                throw new KlumValidationException(aggregatedErrors);
-            return aggregatedErrors;
-        }
-
-        @Override
-        public void visit(@NotNull String path, @NotNull Object element, @Nullable Object container, @Nullable String nameOfFieldInContainer) {
-            KlumInstanceProxy proxy = KlumInstanceProxy.getProxyFor(element);
-            if (proxy.getManualValidation()) return;
-
-            KlumValidationResult result = Validator.lenientValidate(element, path);
-            if (result.getMaxLevel().equalOrWorseThan(Validate.Level.INFO)) {
-                aggregatedErrors.add(result);
-                currentMaxLevel = currentMaxLevel.combine(result.getMaxLevel());
-            }
-        }
-
+        Validator.lenientValidate(element, path);
     }
 }
