+++ klum-ast-runtime/src/main/java/com/blackbuild/klum/ast/util/Validator.java
@@ -26,6 +26,7 @@ package com.blackbuild.klum.ast.util;
 import com.blackbuild.annodocimal.annotations.AnnoDoc;
 import com.blackbuild.groovy.configdsl.transform.Owner;
 import com.blackbuild.groovy.configdsl.transform.Validate;
+import com.blackbuild.klum.ast.util.layer3.StructureUtil;
 import groovy.lang.Closure;
 import org.codehaus.groovy.runtime.InvokerHelper;
 
@@ -111,7 +112,8 @@ public class Validator {
      * @throws KlumValidationException if validation errors are found
      */
     public static List<KlumValidationResult> validateStructure(Object instance, Validate.Level failLevel) throws KlumValidationException {
-        return new ValidationPhase.Visitor().executeOn(instance, failLevel);
+        StructureUtil.visit(instance, new ValidationPhase());
+        return new VerifyPhase.Visitor().executeOn(instance, failLevel);
     }
 
     /**
